moving all checks to a separate file
(issue #1204)
